refactor: extract booking logic into custom hook
- Move booking logic from booking-screen.tsx to useBookingScreen hook
- Create Stylist type and update Service interface
- Streamline booking-screen.tsx by consuming the new hook
